+++ R/agg_pipelines.R
@@ -281,19 +281,10 @@ get_pipeline <- function(countries = "all",
     ####################
 
     targets::tar_target_raw(
-      "ETA_prep",
+      "AggPFU",
       substitute(PSUT_Re_all_Gr_all_Chop_all_St_pfd %>%
-                   add_grossnet_column(countries = CountriesContinentsWorld,
-                                       years = Years)),
-      pattern = quote(cross(CountriesContinentsWorld, Years))
-    ),
-
-
-    targets::tar_target_raw(
-      "ETA_pfd",
-      substitute(ETA_prep %>%
-                   calculate_pfd_efficiencies(countries = CountriesContinentsWorld,
-                                              years = Years)),
+                   calculate_pfu_aggregates(countries = CountriesContinentsWorld,
+                                      years = Years)),
       pattern = quote(cross(CountriesContinentsWorld, Years))
     ),
 
@@ -303,11 +294,11 @@ get_pipeline <- function(countries = "all",
     ####################
 
     targets::tar_target_raw(
-      "ETA_pfu",
-      substitute(ETA_pfd %>%
+      "EtaPFU",
+      # substitute(ETA_pfd %>%
+      substitute(AggPFU %>%
                    calculate_pfu_efficiencies(countries = CountriesContinentsWorld,
                                               years = Years)),
-      # pattern = quote(map(CountriesContinentsWorld))
       pattern = quote(cross(CountriesContinentsWorld, Years))
     ),
 
@@ -318,9 +309,9 @@ get_pipeline <- function(countries = "all",
 
     # Pin the ETA_pfu data frame
     targets::tar_target_raw(
-      "ReleaseETApfu",
+      "ReleaseEtaPFU",
       quote(PFUDatabase::release_target(pipeline_releases_folder = PinboardFolder,
-                                        targ = ETA_pfu,
+                                        targ = EtaPFU,
                                         targ_name = "eta_pfu",
                                         release = Release))),
 
@@ -330,12 +321,12 @@ get_pipeline <- function(countries = "all",
       quote(PFUDatabase::stash_cache(pipeline_caches_folder = PipelineCachesFolder,
                                      cache_folder = "_targets",
                                      file_prefix = "pfu_agg_pipeline_cache_",
-                                     dependency = ETA_pfu,
+                                     dependency = EtaPFU,
                                      release = Release))),
 
     # Write a csv file of efficiencies
     targets::tar_target_raw(
-      "ReleaseETApfuCSV",
+      "ReleaseEtaPFUCSV",
       quote(PFUDatabase::release_target(pipeline_releases_folder = PinboardFolder,
                                         targ = ETA_pfu,
                                         targ_name = "eta_pfu_csv",